refactor: Move Depends() functions to dependencies module. Move ERROR_CODES to util module.
