+++ app/auth.py
@@ -1,15 +1,12 @@
 from typing import Dict, List, Optional
 
 import requests
-from fastapi import Depends, HTTPException, Request, status
+from fastapi import HTTPException, Request, status
 from fastapi.security import HTTPAuthorizationCredentials, HTTPBearer
 from jose import JWTError, jwk, jwt
 from jose.utils import base64url_decode
 from pydantic import BaseModel
-from sqlalchemy.orm import Session
 
-from app.db import get_db
-from app.models import User
 from app.settings import app_settings
 
 JWK = Dict[str, str]
@@ -43,7 +40,7 @@ class verifyTokeClass(HTTPBearer):
 
     def load_keys(self):
         if self.kid_to_jwk is None:
-            jwks = get_public_keys()
+            jwks = _get_public_keys()
             self.kid_to_jwk = {jwk["kid"]: jwk for jwk in jwks.keys}
 
     def verify_jwk_token(self, jwt_credentials: JWTAuthorizationCredentials) -> bool:
@@ -114,7 +111,7 @@ class verifyTokeClass(HTTPBearer):
 JsonPublicKeys = None
 
 
-def get_public_keys():
+def _get_public_keys():
     """
     Retrieves the public keys from the well-known JWKS (JSON Web Key Set) endpoint of AWS Cognito.
 
@@ -133,46 +130,3 @@ def get_public_keys():
             ).json()
         )
     return JsonPublicKeys
-
-
-async def get_auth_credentials(request: Request):
-    return await verifyTokeClass().__call__(request)
-
-
-async def get_current_user(
-    credentials: JWTAuthorizationCredentials = Depends(get_auth_credentials),
-) -> str:
-    """
-    Extracts the username of the current user from the provided JWT credentials.
-
-    :param credentials: JWT credentials provided by the user. Defaults to Depends(get_auth_credentials).
-    :type credentials: JWTAuthorizationCredentials
-    :raises HTTPException: If the username key is missing in the JWT claims.
-    :return: The username of the current user.
-    :rtype: str
-    """
-    try:
-        return credentials.claims["username"]
-    except KeyError:
-        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail="Username missing")
-
-
-async def get_user(
-    credentials: JWTAuthorizationCredentials = Depends(get_auth_credentials),
-    session: Session = Depends(get_db),
-) -> str:
-    """
-    Retrieves the user from the database using the username extracted from the provided JWT credentials.
-
-    :param credentials: JWT credentials provided by the user. Defaults to Depends(get_auth_credentials).
-    :type credentials: JWTAuthorizationCredentials
-    :param session: Database session to execute the query. Defaults to Depends(get_db).
-    :type session: Session
-    :raises HTTPException: If the user does not exist in the database.
-    :return: The user object retrieved from the database.
-    :rtype: User
-    """
-    try:
-        return User.first_by(session, "external_id", credentials.claims["username"])
-    except KeyError:
-        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail="User not found")
