+++ tests/protected_routes/users_test.py
@@ -1,7 +1,8 @@
 from fastapi import APIRouter, Depends
 from sqlalchemy.orm import Session
 
-from app.aws import CognitoClient, get_cognito_client
+from app import dependencies
+from app.aws import CognitoClient
 from app.db import get_db
 from app.models import User
 
@@ -15,7 +16,7 @@ router = APIRouter()
 async def create_test_user_headers(
     payload: User,
     session: Session = Depends(get_db),
-    client: CognitoClient = Depends(get_cognito_client),
+    client: CognitoClient = Depends(dependencies.get_cognito_client),
 ):
     user = User(**payload.dict())
